Cleanup layout files and split up
